Remove ECS code and EnTT dependency
The ECS system was partially implemented but only used for player state
management. This simplifies the codebase by:

- Removing the EnTT library dependency
- Deleting ECS files (World.h, Components.h, Systems.h, LightIntegration.h)
- Replacing World with a simple PlayerState struct containing transform,
  movement settings, and grounded state
- Updating Application to use PlayerState directly instead of ECS queries

The unused NPC/AI components, light entity system, and patrol/health
systems have been removed entirely.
